Rename storage -> models
- Rename module (storage -> models)
- Split daatabase models into different source files
- Move db singleton instance to models module
