+++ app/main/models/__init__.py
@@ -1,0 +1,8 @@
+from flask_sqlalchemy import SQLAlchemy
+
+db = SQLAlchemy()
+
+
+def get_db_instance():
+    global db
+    return db
